refactor send_transaction() and send_raw_transaction()
